WIP [4/4] Remove coroutines from TransitivePathImpl.h
Note: The TransitiveHull generator has a different type and yield type from the other generators that are migrated. I'm not sure how to proceed here. It could be that there is a simple conversion that can be done, or it could be that there is a very different concept to apply for migrating this generator