WIP Attempt to solve problem in previous commit, does not work yet
Note: Now the problem is that the templated generator is used with two different types. One is "const Set", so it requires a const_iterator for iteration. The other is "span<Id>", and since span doesn't have a const_iterator until C++23, this will not compile with the current return types.